[nrf temphack] boards: qemu_cortex_r5: Make the board non-default
Remove the `default` key that is set for the board, so that it is
not picked automatically by the sanitycheck script.
This hack is needed until the zephyr-sdk is updated to 0.11.0,
as this board currently requires testing on QEMU not available in
version 0.10.3. See commit 59cb2acd9ec1d03a0a3401fbccb7ab608649979b.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
(cherry picked from commit a32d8dbbef271f84fd9ce15214db00d714e8fc42)